+++ src/sciform/__init__.py
@@ -1,7 +1,4 @@
-__version__ = "0.17.1"
-
-# TODO: PDG sig figs
-# TODO: get exponent based on larger of val/unc
+__version__ = "0.18.0"
 
 from sciform.formatter import Formatter
 from sciform.float_formatting import sfloat, vufloat
@@ -11,13 +8,15 @@ from sciform.format_options import (
     global_add_ppth_form, global_reset_si_prefixes,
     global_reset_iec_prefixes, global_reset_parts_per_forms,
     print_global_defaults, GlobalDefaultsContext)
-from sciform.modes import (FillMode, SignMode, GroupingSeparator, RoundMode,
-                           ExpMode)
 
-__all__ = ['__version__', 'Formatter', 'sfloat', 'set_global_defaults',
-           'reset_global_defaults', 'global_add_c_prefix',
-           'global_add_small_si_prefixes', 'global_reset_si_prefixes',
-           'global_add_ppth_form', 'global_reset_iec_prefixes',
-           'global_reset_parts_per_forms', 'print_global_defaults',
-           'GlobalDefaultsContext', 'FillMode', 'SignMode',
-           'GroupingSeparator', 'RoundMode', 'ExpMode', 'vufloat']
+from sciform.modes import (FillMode, SignMode, GroupingSeparator, RoundMode,
+                           ExpMode, AutoExp, AutoPrec)
+
+__all__ = ['__version__', 'Formatter', 'sfloat', 'vufloat',
+           'set_global_defaults', 'reset_global_defaults',
+           'global_add_c_prefix', 'global_add_small_si_prefixes',
+           'global_reset_si_prefixes', 'global_add_ppth_form',
+           'global_reset_iec_prefixes', 'global_reset_parts_per_forms',
+           'print_global_defaults', 'GlobalDefaultsContext', 'FillMode',
+           'SignMode', 'GroupingSeparator', 'RoundMode', 'ExpMode', 'AutoExp',
+           'AutoPrec']